third_party/blink/web_tests/external/wpt/url/urlsearchparams-delete.any.worker-expected.txt

This is a testharness.js-based test.
[FAIL] Changing the query of a URL with an opaque path with trailing spaces
  assert_equals: expected "space   %20" but got "space    "
[FAIL] Changing the query of a URL with an opaque path with trailing spaces and a fragment
  assert_equals: expected "space   %20" but got "space    "
Harness: the test ran to completion.
